Refactor match logic into reusable controller and classic mode
